Round the dropdown vertical coordinate to device pixels
In Safari the item text is sometimes rendered 1 device pixel outside the dropdown when using long vertical-scrolled iron-list inside an iron-dropdown that is positioned to a position target with non-integer offsets.

Issue demo in JSFiddle: http://jsfiddle.net/3mprhn5p/2/

More noticeable on non-retina screens.

Generally this seems to be a Safari rendering issue. Occurs when CSS transforms are applied on the element containing text, while the element’s parent is positioned on non-integer pixel coordinates. The rendering issue could be reproduced without Polymer Elements and Web Components as well, see this demo: http://jsfiddle.net/fghtdh7r/6/

The proposed fix makes sure that the vertical coordinate of a dropdown is rounded to a whole number of device pixels. This effectively prevents rounding issues with Safari rendering engine.